+++ src/app/pages/fna-bram/inputs/management-infrastructure/location-business/location-business.component.ts
@@ -43,12 +43,18 @@ export class LocationBusinessComponent implements OnInit {
   }
 
   delete(id) {
-    this.managementInfrastructureService.deleteManagementInfrastructureById(id).toPromise().then(async (res: any) => {
-      if (res) {
-        this.getAll();
-        this.alertService.activate('This record was deleted', 'Success Message');
+    this.alertService.activate('Are you sure you want to delete?', 'Warning Message').then(result => {
+      if (result) {
+        this.managementInfrastructureService.deleteManagementInfrastructureById(id).toPromise().then(async (res: any) => {
+          if (res) {
+            this.getAll();
+            this.alertService.activate('This record was deleted', 'Success Message').then(result => {
+           
+            });
+          }
+        })
       }
-    })
+    });
   }
 
   async getAll() {
