+++ src/app/pages/fna-bram/inputs/marketing-sales/marketing-sales-others/marketing-sales-others.component.ts
@@ -45,12 +45,18 @@ export class MarketingSalesOthersComponent implements OnInit {
   }
 
   delete(id) {
-    this.marketingSalesService.deleteMarketingSalesById(id).toPromise().then(async (res: any) => {
-      if (res) {
-        this.getAll();
-        this.alertService.activate('This record was deleted', 'Success Message');
+    this.alertService.activate('Are you sure you want to delete?', 'Warning Message').then(result => {
+      if (result) {
+        this.marketingSalesService.deleteMarketingSalesById(id).toPromise().then(async (res: any) => {
+          if (res) {
+            this.getAll();
+            this.alertService.activate('This record was deleted', 'Success Message').then(result => {
+           
+            });
+          }
+        })
       }
-    })
+    });
   }
 
   async getAll() {
