+++ src/app/pages/fna-bram/inputs/operation/operation-others/operation-others.component.ts
@@ -45,12 +45,18 @@ export class OperationOthersComponent implements OnInit {
   }
 
   async delete(id) {
-    this.operationService.deleteOperationById(id).toPromise().then(async (res: any) => {
-      if (res) {
-        await this.getAll();
-        this.alertService.activate('This record was deleted', 'Success Message');
+    this.alertService.activate('Are you sure you want to delete?', 'Warning Message').then(result => {
+      if (result) {
+        this.operationService.deleteOperationById(id).toPromise().then(async (res: any) => {
+          if (res) {
+            await this.getAll();
+            this.alertService.activate('This record was deleted', 'Success Message').then(result => {
+           
+            });
+          }
+        })
       }
-    })
+    });
   }
 
   async getAll() {
