+++ src/app/pages/fna-bram/inputs/technology/hardware/hardware.component.ts
@@ -46,12 +46,18 @@ export class HardwareComponent implements OnInit {
 
 
   delete(id) {
-    this.technologyService.deleteTechnologyById(id).toPromise().then(async (res: any) => {
-      if (res) {
-        this.getAll();
-        this.alertService.activate('This record was deleted', 'Success Message');
+    this.alertService.activate('Are you sure you want to delete?', 'Warning Message').then(result => {
+      if (result) {
+        this.technologyService.deleteTechnologyById(id).toPromise().then(async (res: any) => {
+          if (res) {
+            this.getAll();
+            this.alertService.activate('This record was deleted', 'Success Message').then(result => {
+           
+            });
+          }
+        })
       }
-    })
+    });
   }
 
   async getAll() {
