+++ src/app/pages/product-form/product-form.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectorRef, Component, OnDestroy, OnInit, ViewChild,NgZone } from '@angular/core';
+import { ChangeDetectorRef, Component, OnDestroy, OnInit, ViewChild, NgZone } from '@angular/core';
 import { of, Subscription } from 'rxjs';
 import { DatePipe, DecimalPipe, Location } from '@angular/common';
 // import { uuid } from 'uuid';
@@ -66,7 +66,13 @@ export class ProductFormComponent implements OnInit, OnDestroy {
   editData: any
   creatingCustomer: Customer = {}
   creatingLeadId: number = 0
-  constructor(private router: Router, private location: Location, private cdRef: ChangeDetectorRef, private modalService: NgbModal, private prodService: ProductDataService, private globalFun: GlobalFunctionService, private auth: AuthService, private pageDataService: PageDataService, private addonQuo: AddOnQuoService, private coverageQuo: CoverageQuoService, private alert: AlertService, private downloadService: AttachmentDownloadService, private masterServer: MasterDataService, private numberPipe: DecimalPipe, private datePipe: DatePipe,private ngZone:NgZone) { }
+  constructor(private router: Router, private location: Location, private cdRef: ChangeDetectorRef,
+    private modalService: NgbModal, private prodService: ProductDataService,
+    private globalFun: GlobalFunctionService, private auth: AuthService,
+    private pageDataService: PageDataService, private addonQuo: AddOnQuoService,
+    private coverageQuo: CoverageQuoService, private alert: AlertService,
+    private downloadService: AttachmentDownloadService, private masterServer: MasterDataService,
+    private numberPipe: DecimalPipe, private datePipe: DatePipe, private ngZone: NgZone,) { }
 
   async ngOnInit() {
     if ((this.prodService.type == 'policy' && this.prodService.createingProdRef)) {
@@ -698,13 +704,20 @@ export class ProductFormComponent implements OnInit, OnDestroy {
   }
 
   deleteData(index, data) {
-    let activeForm = this.formData[this.activePage]
-    this.pageDataService.deleteData(activeForm.tableName, data.refId, activeForm.id).toPromise().then((res) => {
-      if (res) {
-        this.tempData[activeForm.tableName + activeForm.id].splice(index, 1)
-        this.cdRef.detectChanges();
+    this.alert.activate('Are you sure you want to delete?', 'Warning Message').then(result => {
+      if (result) {
+        let activeForm = this.formData[this.activePage]
+        this.pageDataService.deleteData(activeForm.tableName, data.refId, activeForm.id).toPromise().then((res) => {
+          if (res) {
+            this.tempData[activeForm.tableName + activeForm.id].splice(index, 1)
+            this.cdRef.detectChanges();
+            this.alert.activate('This record was deleted', 'Success Message').then(result => {
+           
+            });
+          }
+        })
       }
-    })
+    });   
   }
